# generated/api/PawInHand/rest/model_volunteer_registration.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * PawInHand API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 */

package modules

type VolunteerRegistration struct {
	UserId string `json:"userId"`

	Preferences string `json:"preferences,omitempty"`
}

// AssertVolunteerRegistrationRequired checks if the required fields are not zero-ed
func AssertVolunteerRegistrationRequired(obj VolunteerRegistration) error {
	elements := map[string]interface{}{
		"userId": obj.UserId,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertVolunteerRegistrationConstraints checks if the values respects the defined constraints
func AssertVolunteerRegistrationConstraints(obj VolunteerRegistration) error {
	return nil
}