# generated/api/PawInHand/rest/model_volunteer_work.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * PawInHand API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 */

package modules

type VolunteerWork struct {
	Id string `json:"id,omitempty"`

	Title string `json:"title"`

	Description string `json:"description"`

	Date string `json:"date,omitempty"`

	Location VolunteerWorkLocation `json:"location,omitempty"`
}

// AssertVolunteerWorkRequired checks if the required fields are not zero-ed
func AssertVolunteerWorkRequired(obj VolunteerWork) error {
	elements := map[string]interface{}{
		"title":       obj.Title,
		"description": obj.Description,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	if err := AssertVolunteerWorkLocationRequired(obj.Location); err != nil {
		return err
	}
	return nil
}

// AssertVolunteerWorkConstraints checks if the values respects the defined constraints
func AssertVolunteerWorkConstraints(obj VolunteerWork) error {
	if err := AssertVolunteerWorkLocationConstraints(obj.Location); err != nil {
		return err
	}
	return nil
}